Revert "Change blocknumber as the time to operate in blockchain"
